updated documentation
