+++ README.md
@@ -1,23 +1,36 @@
 ![CI/CD](https://github.com/gkrumbach07/thoth-search/workflows/CI/CD/badge.svg)
 
-# Thoth Search
-
 This is a React powered web app for displaying Python package metadata and insights using Thoth Advise.
 
 ## Website
 You can check of the latest build here: [gkrumbach07.github.io/thoth-search](https://gkrumbach07.github.io/thoth-search)
 
-## Code Journal
+## 🗄️ Project Structure
 
-Check [this](https://docs.google.com/document/d/1YwA_YZ--BYtaYBF0I40of15vW5E1xNnEuaraTtYZnTU/edit?usp=sharing) out for most up to date development progress.
+Most of the code lives in the `src` folder and looks like this:
 
-## New Features
-
-If you want to see a feature added, make an issue and label it as a new feature.
-
-## Bugs
-
-If you find a bug in the latest, make an issue and label it as a bug.
-
-`yarn`
-`yarn start`
+```text
+src
+|
++-- assets            # assets folder can contain all the static files such as images, fonts, etc.
+|
++-- components        # shared components used across the entire application
+|
++-- config            # all the global configuration, env variables etc. get exported from here and used in the app
+|
++-- features          # feature based modules
+|
++-- hooks             # shared hooks used across the entire application
+|
++-- lib               # re-exporting different libraries preconfigured for the application
+|
++-- providers         # all of the application providers
+|
++-- routes            # routes configuration
+|
++-- stores            # global state stores
+|
++-- test              # test utilities and mock server
+|
++-- utils             # shared utility functions
+```
