+++ src/features/advise/components/AdviseHeader/AdviseHeader.md
@@ -1,10 +1,12 @@
 ```jsx
 <AdviseHeader
-    adviseID={"advise-adviser-211112223258-38af1a4746733b53"}
-    adviseReport={{
-            status: {
-                finished_at:"2021-10-05T14:48:00.000Z"
-            }
-        }}
+   adviseDocument={{
+       status: {
+           finished_at:"2021-10-05T14:48:00.000Z"
+       },
+       metadata: {
+           document_id: "adviser-211112223258-38af1a4746733b53",
+       }
+   }}
 />
 ```
