--- src/features/advise/components/SelectedPackage/NetworkGraphView/Popup/Popup.md
@@ -1,7 +1,0 @@
-```jsx
-import {Graph} from "utils/Graph"
-const graph = new Graph();
-const node1 = graph.addVertex("1", {label: "node 1", id: "1"});
-
-<Popup node={node1} />
-```
