--- src/features/advise/components/SelectedPackage/NetworkGraphView/Popup/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Popup";
