--- src/features/advise/components/SelectedPackage/NetworkGraphView/index.js
@@ -1,1 +1,0 @@
-export { default } from "./NetworkGraphView";
