--- src/features/advise/components/SelectedPackage/index.js
@@ -1,1 +1,0 @@
-export * from "./SelectedPackage";
