+++ src/features/advise/components/index.js
@@ -4,5 +4,4 @@ export * from "./CustomCardAction"
 export * from "./DependenciesMetric"
 export * from "./LicenseMetric"
 export * from "./AdviseTableView"
-export * from "./SelectedPackage"
 export * from "./LockfileView"
