+++ src/features/advise/features/components/NetworkGraphView/index.js
@@ -1,0 +1,1 @@
+export { default } from "./NetworkGraphView";
