+++ src/features/advise/features/components/Popup/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Popup";
