+++ src/features/advise/features/components/SelectedPackage/SelectedPackage.md
@@ -1,0 +1,8 @@
+It consists of multiple sections including:
+- Package metadata display (summary, version, license, ...)
+- Thoth justifications for package
+- Calculated reasons for removal, addition, or version changes
+- Where the package originated from (its dependent tree)
+- A graph of the dependent paths back to root.
+
+_this section needs to be elaborated on_
