+++ src/features/features.md
@@ -1,0 +1,45 @@
+In order to scale the application in the easiest and most maintainable way, keep most of the code inside the `features` folder, which should contain different feature-based things. Every `feature` folder should contain domain specific code for a given feature. This will allow you to keep functionalities scoped to a feature and not mix its declarations with shared things. This is much easier to maintain than a flat folder structure with many files.
+
+A feature could have the following structure:
+
+```text
+src/features/awesome-feature
+|
++-- api         # exported API request declarations and api hooks related to a specific feature
+|
++-- assets      # assets folder can contain all the static files for a specific feature
+|
++-- components  # components scoped to a specific feature
+|
++-- features    # sub-features scoped to a specific feature
+|
++-- hooks       # hooks scoped to a specific feature
+|
++-- routes      # route components for a specific feature pages
+|
++-- stores      # state stores for a specific feature
+|
++-- types       # typescript types for TS specific feature domain
+|
++-- utils       # utility functions for a specific feature
+|
++-- index.ts    # entry point for the feature, it should serve as the public API of the given feature and exports everything that should be used outside the feature
+```
+
+A feature folder could also contain other features (if used only within the parent feature) or be kept separated, it's a matter of preference.
+
+Everything from a feature should be exported from the `index.js` file which behaves as the public API of the feature.
+
+You should import stuff from other features only by using:
+
+``` js static
+import {AwesomeComponent} from "features/awesome-feature"
+```
+
+and not
+
+``` js static
+import {AwesomeComponent} from "features/awesome-feature/components/AwesomeComponent"
+```
+
+This structure is taken from [github.com/alan2207/bulletproof-react](https://github.com/alan2207/bulletproof-react)
