--- src/features/misc/routes/Landing.js
@@ -1,8 +1,0 @@
-
-export const Landing = () => {
-    return (
-        <>
-
-        </>
-    );
-};
