+++ src/features/misc/routes/index.js
@@ -1,2 +1,1 @@
 export * from "./NotFound"
-export * from "./Landing"
