+++ src/features/package/package.md
@@ -1,0 +1,4 @@
+The package feature page displays specific metadata and metrics for a single python package.
+
+### 📄 Pages
+There is one page that consists of a metadata header and a grid of metrics.
